refactor(tesorería): las métricas de tarjeta viven en su propio servicio
`tesoreriaV6Metrics` se iba a 924 líneas y cruzaba el umbral de 800 del
trinquete de salud (38 → 39). Sacarlas no es solo bajar el contador: una
tarjeta no es una cuenta —no tiene saldo, tiene un ciclo, y su dinero sale en
el recibo, no el día de la compra—, así que sus métricas no tenían por qué
vivir en el fichero de las cuentas.

`soloFecha`, `enRango` y `redondear` pasan a exportarse: los necesita el módulo
nuevo y duplicarlos sería tener dos definiciones de "qué día es esto".

Sin cambio de comportamiento · los 366 tests siguen en verde.
